Refactor data

+++ oxidation/libparsec_python/src/local_manifest.rs
@@ -1240,7 +1240,7 @@ impl LocalUserManifest {
             [base: UserManifest, "base"],
             [need_sync: bool, "need_sync"],
             [updated, "updated", py_to_rs_datetime],
-            [last_processed_message: u32, "last_processed_message"],
+            [last_processed_message: u64, "last_processed_message"],
             [workspaces: Vec<WorkspaceEntry>, "workspaces"],
             [speculative: bool, "speculative"],
         );
@@ -1262,7 +1262,7 @@ impl LocalUserManifest {
             [base: UserManifest, "base"],
             [need_sync: bool, "need_sync"],
             [updated, "updated", py_to_rs_datetime],
-            [last_processed_message: u32, "last_processed_message"],
+            [last_processed_message: u64, "last_processed_message"],
             [workspaces: Vec<WorkspaceEntry>, "workspaces"],
             [speculative: bool, "speculative"],
         );
@@ -1437,7 +1437,7 @@ impl LocalUserManifest {
     }
 
     #[getter]
-    fn last_processed_message(&self) -> PyResult<u32> {
+    fn last_processed_message(&self) -> PyResult<u64> {
         Ok(self.0.last_processed_message)
     }
